fix some bugs

+++ install-kubectl-in-cluster.sh
@@ -1,27 +1,34 @@
 #!/usr/bin/env bash
 
-# Point to the internal API server hostname
-APISERVER=https://kubernetes.default.svc
+if ! command -v kubectl &> /dev/null
+then
+    echo "kubectl could not be found. Installing it..."
 
-# Path to ServiceAccount token
-SERVICEACCOUNT=/var/run/secrets/kubernetes.io/serviceaccount
+    # Point to the internal API server hostname
+    APISERVER=https://kubernetes.default.svc
 
-# Read this Pod's namespace
-NAMESPACE=$(cat ${SERVICEACCOUNT}/namespace)
+    # Path to ServiceAccount token
+    SERVICEACCOUNT=/var/run/secrets/kubernetes.io/serviceaccount
 
-# Read the ServiceAccount bearer token
-TOKEN=$(cat ${SERVICEACCOUNT}/token)
+    # Read this Pod's namespace
+    NAMESPACE=$(cat ${SERVICEACCOUNT}/namespace)
 
-# Reference the internal certificate authority (CA)
-CACERT=${SERVICEACCOUNT}/ca.crt
+    # Read the ServiceAccount bearer token
+    TOKEN=$(cat ${SERVICEACCOUNT}/token)
 
-# Explore the API with TOKEN
-VERSION_DETAILS=$(curl -Ls --cacert ${CACERT} --header "Authorization: Bearer ${TOKEN}" -X GET ${APISERVER}/version)
-KUBECTL_PLATFORM=$(echo $VERSION_DETAILS | jq -rc '.platform')
-KUBECTL_VERSION=$(echo $VERSION_DETAILS | jq -rc '.gitVersion')
+    # Reference the internal certificate authority (CA)
+    CACERT=${SERVICEACCOUNT}/ca.crt
 
-curl -L "https://dl.k8s.io/release/${KUBECTL_VERSION}/bin/${KUBECTL_PLATFORM}/kubectl" -o /usr/local/bin/kubectl \
-  && curl -L "https://dl.k8s.io/${KUBECTL_VERSION}/bin/${KUBECTL_PLATFORM}/kubectl.sha256" -o /tmp/kubectl.sha256 \
-  && echo "$(cat /tmp/kubectl.sha256)  /usr/local/bin/kubectl" | sha256sum -c \
-  && rm /tmp/kubectl.sha256 \
-  && chmod +x /usr/local/bin/kubectl
+    # Explore the API with TOKEN
+    VERSION_DETAILS=$(curl -Ls --cacert ${CACERT} --header "Authorization: Bearer ${TOKEN}" -X GET ${APISERVER}/version)
+    KUBECTL_PLATFORM=$(echo $VERSION_DETAILS | jq -rc '.platform')
+    KUBECTL_VERSION=$(echo $VERSION_DETAILS | jq -rc '.gitVersion')
+
+    curl -L "https://dl.k8s.io/release/${KUBECTL_VERSION}/bin/${KUBECTL_PLATFORM}/kubectl" -o /usr/local/bin/kubectl \
+      && curl -L "https://dl.k8s.io/${KUBECTL_VERSION}/bin/${KUBECTL_PLATFORM}/kubectl.sha256" -o /tmp/kubectl.sha256 \
+      && echo "$(cat /tmp/kubectl.sha256)  /usr/local/bin/kubectl" | sha256sum -c \
+      && rm /tmp/kubectl.sha256 \
+      && chmod +x /usr/local/bin/kubectl
+    
+    echo "kubectl installed"
+fi
